Refactor project into input processing output
