+++ .github/workflows/test.yml
@@ -19,5 +19,5 @@ jobs:
           cache: pip
       - run: python -m pip install --upgrade pip
       - run: python -m pip install -r requirements.txt
-      - run: python -m py_compile photogrammetry.py config.py video_pipeline.py tests/test_photogrammetry.py tests/test_video_pipeline.py
+      - run: python -m py_compile main.py processing/*.py tests/*.py
       - run: python -m unittest discover -s tests -v
