+++ processing/__init__.py
@@ -1,0 +1,1 @@
+"""Processing modules for AutoPhotogrammetry."""
