+++ requirements.txt
@@ -3,4 +3,3 @@ numpy>=1.24,<3
 Pillow>=10,<12
 requests>=2.31,<3
 scikit-image>=0.22,<1
-scikit-learn>=1.4,<2
