+++ tests/test_photogrammetry.py
@@ -5,10 +5,11 @@ import unittest
 from pathlib import Path
 from unittest.mock import patch
 
-from photogrammetry import (
+from processing.photogrammetry import (
     BackendConfig,
     BackendConfigurationError,
     build_command,
+    load_backend_configs,
     resolve_executable,
     run_backend,
 )
@@ -19,21 +20,70 @@ class PhotogrammetryRunnerTests(unittest.TestCase):
         image_dir = Path("input images")
         output_dir = Path("output models")
         commands = {
-            "meshroom": ["tool", "--input", str(image_dir), "--output", str(output_dir)],
-            "visualsfm": ["tool", str(image_dir), str(output_dir)],
+            "meshroom": [
+                "tool",
+                "--input",
+                str(image_dir),
+                "--output",
+                str(output_dir),
+            ],
+            "visualsfm": [
+                "tool",
+                str(image_dir),
+                str(output_dir),
+            ],
             "colmap": [
-                "tool", "automatic_reconstructor", "--image_path", str(image_dir),
-                "--workspace_path", str(output_dir),
+                "tool",
+                "automatic_reconstructor",
+                "--image_path",
+                str(image_dir),
+                "--workspace_path",
+                str(output_dir),
             ],
         }
         for backend, expected in commands.items():
             with self.subTest(backend=backend):
-                self.assertEqual(build_command(backend, "tool", image_dir, output_dir), expected)
+                self.assertEqual(
+                    build_command(
+                        backend,
+                        "tool",
+                        image_dir,
+                        output_dir,
+                    ),
+                    expected,
+                )
 
     def test_missing_executable_fails_with_configuration_guidance(self):
-        with patch("photogrammetry.shutil.which", return_value=None):
-            with self.assertRaisesRegex(BackendConfigurationError, "AUTOPHOTOGRAMMETRY_COLMAP_EXECUTABLE"):
+        with patch(
+            "processing.photogrammetry.shutil.which",
+            return_value=None,
+        ):
+            with self.assertRaisesRegex(
+                BackendConfigurationError,
+                "AUTOPHOTOGRAMMETRY_COLMAP_EXECUTABLE",
+            ):
                 resolve_executable("colmap", env={})
+
+    def test_config_loading_stays_with_photogrammetry(self):
+        with tempfile.TemporaryDirectory() as tmp:
+            path = Path(tmp) / "backends.json"
+            path.write_text(
+                json.dumps(
+                    {
+                        "colmap": {
+                            "executable": "/usr/bin/colmap",
+                            "extra_args": ["--quality", "medium"],
+                        }
+                    }
+                ),
+                encoding="utf-8",
+            )
+            config = load_backend_configs(path)["colmap"]
+            self.assertEqual(config.executable, "/usr/bin/colmap")
+            self.assertEqual(
+                config.extra_args,
+                ("--quality", "medium"),
+            )
 
     def test_run_writes_separate_manifest_logs_and_artifacts(self):
         with tempfile.TemporaryDirectory() as tmp:
@@ -54,7 +104,9 @@ class PhotogrammetryRunnerTests(unittest.TestCase):
                 "print('completed')\n",
                 encoding="utf-8",
             )
-            executable.chmod(executable.stat().st_mode | stat.S_IXUSR)
+            executable.chmod(
+                executable.stat().st_mode | stat.S_IXUSR
+            )
 
             result = run_backend(
                 "meshroom",
@@ -62,14 +114,24 @@ class PhotogrammetryRunnerTests(unittest.TestCase):
                 output_root,
                 BackendConfig(executable=str(executable)),
             )
-            manifest = json.loads(Path(result.manifest_path).read_text(encoding="utf-8"))
+            manifest = json.loads(
+                Path(result.manifest_path).read_text(
+                    encoding="utf-8"
+                )
+            )
             self.assertEqual(result.return_code, 0)
             self.assertEqual(manifest["backend"], "meshroom")
             self.assertEqual(manifest["version"], "fake 1.0")
-            self.assertIn("model/result.txt", manifest["artifacts"])
-            self.assertEqual(Path(result.stdout_log).read_text(encoding="utf-8"), "completed\n")
-            self.assertTrue(any("input images" in arg for arg in result.command))
-            self.assertTrue(Path(result.run_dir).is_relative_to(output_root / "meshroom"))
+            self.assertIn(
+                "model/result.txt",
+                manifest["artifacts"],
+            )
+            self.assertEqual(
+                Path(result.stdout_log).read_text(
+                    encoding="utf-8"
+                ),
+                "completed\n",
+            )
 
 
 if __name__ == "__main__":
